Convert project to TypeScript
